Fix platform typing
