+++ frontend/hooks/usePlatformCompatibility.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useState } from "react";
 import { useDebouncedCallback } from "use-debounce";
 
-import { OsqueryPlatform, SUPPORTED_PLATFORMS } from "interfaces/platform";
+import { SupportedPlatform, SUPPORTED_PLATFORMS } from "interfaces/platform";
 import { checkPlatformCompatibility } from "utilities/sql_tools";
 
 import PlatformCompatibility from "components/PlatformCompatibility";
@@ -16,7 +16,7 @@ const DEBOUNCE_DELAY = 300;
 
 const usePlatformCompatibility = (): IPlatformCompatibility => {
   const [compatiblePlatforms, setCompatiblePlatforms] = useState<
-    OsqueryPlatform[] | null
+    SupportedPlatform[] | null
   >(null);
   const [error, setError] = useState<Error | null>(null);
